--- src/by/it/akhmelev/jd03_01/classwork/Ex_01_Connection.java
@@ -1,32 +1,0 @@
-package by.it.akhmelev.jd03_01.classwork;
-
-import com.mysql.fabric.jdbc.FabricMySQLDriver;
-
-import java.sql.Connection;
-import java.sql.Driver;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-
-public class Ex_01_Connection {
-
-
-    public static void main(String[ ] args) {
-        try {
-            Driver driver = new FabricMySQLDriver();
-            DriverManager.registerDriver(driver);
-        } catch (SQLException e) {e.printStackTrace();}
-        try (Connection connection=
-                     DriverManager.getConnection
-                             (CN.URL_DB, CN.USER_DB, CN.PASSWORD_DB);){
-                if (!connection.isClosed())
-                    System.out.println("Соединение установлено");
-                connection.close();
-                if (connection.isClosed())
-                    System.out.println("Соединение прервано");
-            }
-            catch (Exception e){
-                e.printStackTrace();
-            }
-
-    }
-}
