--- src/by/it/akhmelev/jd03_01/classwork/Ex_03_Select.java
@@ -1,29 +1,0 @@
-package by.it.akhmelev.jd03_01.classwork;
-
-import com.mysql.fabric.jdbc.FabricMySQLDriver;
-
-import java.sql.*;
-
-public class Ex_03_Select {
-    public static void main(String[] args) {
-        try {
-            Driver driver = new FabricMySQLDriver();
-            DriverManager.registerDriver(driver);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        try (Connection connection =
-                     DriverManager.getConnection
-                             (CN.URL_DB, CN.USER_DB, CN.PASSWORD_DB);
-             Statement statement = connection.createStatement();) {
-            //получаем пользователей
-            ResultSet resultSet = statement.executeQuery("select * from users;");
-            while (resultSet.next()) {
-                String out = resultSet.getString("Email") + ", " + resultSet.getString("Login");
-                System.out.println(out);
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-}
