--- src/by/it/voronovich/JD03_01_Check_here/CN.java
@@ -1,9 +1,0 @@
-package by.it.voronovich.JD03_01_Check_here;
-
-public class CN {
-
-    public static final String URL_DB = "jdbc:mysql://127.0.0.1:2016/voronovich_project"+
-            "?useUnicode=true&amp;characterEncoding=UTF-8";
-    public static final String USER_DB = "root";
-    public static final String PASSWORD_DB = "";
-}
